api/requirements.txt
azure-functions
azure-storage-blob
azure-identity
openai
python-docx
docxtpl
pdfplumber
pydantic
